Move table provider out of lib so we can add in scalar udf without confusion
